server/src/requirements.txt
connexion == 1.1.15
python_dateutil == 2.6.0
setuptools >= 21.0.0
sqlalchemy 
flask-sqlalchemy == 2.1.0
flask-migrate
Werkzeug==0.16.1
flask_testing